Internet based Quickselect implementation

+++ ADA_Assignment2/ADA_Assignment2/Defs.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef double arrayType;
+typedef int arrayType;
